Fix the usage of constant PathParam in handler's request
Dynamic Request that is passed to the handler was having pured `PathParam`
field. Additionally there was a subtle bug in the `Dynamic t DomRequest` was
passed down in the route method. Constant value from the closure was used
to trim the pathInfo causing the loss of Dynamic effect in `PathParam`